[MISC] Rename chopper to hibf.
