+++ src/count_HIBF_kmers_based_on_binning.cpp
@@ -3,7 +3,7 @@
 #include <seqan3/io/sequence_file/input.hpp>
 #include <seqan3/search/views/kmer_hash.hpp>
 
-#include <chopper/build/read_data_file_and_set_high_level_bins.hpp>
+#include <hibf/build/read_data_file_and_set_high_level_bins.hpp>
 
 struct cmd_arguments
 {
