+++ test/api/search/chopper_search_test.cpp
@@ -3,13 +3,13 @@
 #include <fstream>
 #include <sstream>
 
-#include <chopper/build/chopper_build.hpp>
-#include <chopper/build/create_ibfs_from_chopper_pack.hpp>
-#include <chopper/search/pair_hash.hpp>
-#include <chopper/search/chopper_search.hpp>
-#include <chopper/search/search.hpp>
-#include <chopper/search/search_data.hpp>
-#include <chopper/search/sync_out.hpp>
+#include <hibf/build/chopper_build.hpp>
+#include <hibf/build/create_ibfs_from_chopper_pack.hpp>
+#include <hibf/search/pair_hash.hpp>
+#include <hibf/search/chopper_search.hpp>
+#include <hibf/search/search.hpp>
+#include <hibf/search/search_data.hpp>
+#include <hibf/search/sync_out.hpp>
 
 #include "../api_test.hpp"
