+++ test/cli/cli_chopper_pipeline_test.cpp
@@ -10,7 +10,7 @@
 #include <seqan3/utility/views/join_with.hpp>
 #include <seqan3/utility/views/to.hpp>
 
-#include <chopper/detail_bin_prefixes.hpp>
+#include <hibf/detail_bin_prefixes.hpp>
 
 #include "cli_test.hpp"
